requirements.txt
kagglehub==0.3.13
boto3==1.28.57
pandas==1.5.3
sqlalchemy==1.4.52
python-dotenv==1.0.1
pyarrow==12.0.1
requests==2.31.0
psycopg2-binary==2.9.9
minio==7.2.10
pyspark==3.5.1